Restructure post process hooks
