+++ front/lib/upsert_document.ts
@@ -17,10 +17,10 @@ import * as t from "io-ts";
 import { v4 as uuidv4 } from "uuid";
 
 import { Authenticator } from "@app/lib/auth";
+import { getDocumentsPostUpsertHooksToRun } from "@app/lib/documents_post_process_hooks/hooks";
 import logger from "@app/logger/logger";
 import { statsDClient } from "@app/logger/withlogging";
 import { launchRunPostUpsertHooksWorkflow } from "@app/temporal/documents_post_process_hooks/client";
-import { getDocumentsPostUpsertHooksToRun } from "@app/temporal/documents_post_process_hooks/hooks";
 import { launchUpsertDocumentWorkflow } from "@app/temporal/upsert_queue/client";
 
 const { DUST_UPSERT_QUEUE_BUCKET, SERVICE_ACCOUNT } = process.env;
