--- front/temporal/documents_post_process_hooks/hooks/document_tracker/index.ts
@@ -1,2 +1,0 @@
-export { documentTrackerSuggestChangesPostProcessHook } from "@app/temporal/documents_post_process_hooks/hooks/document_tracker/suggest_changes";
-export { documentTrackerUpdateTrackedDocumentsPostProcessHook } from "@app/temporal/documents_post_process_hooks/hooks/document_tracker/update_tracked_documents";
